requirements.txt
Flask
flask_sqlalchemy
requests
python-dotenv
